+++ metadata.rb
@@ -4,10 +4,9 @@ maintainer_email 'david.martin@feedhenry.com'
 license          'MIT License'
 description      'Installs/Configures gitlab-shell'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.4.0'
+version          '0.5.8'
 
-%w(build-essential zlib readline ncurses git openssh redisio xml
-   ruby_build certificate database logrotate apt).each do |cb_depend|
+%w(build-essential zlib readline ncurses git redisio xml ruby_build certificate).each do |cb_depend|
   depends cb_depend
 end